FE: rename emptyAclList constant, fix hook name
